moves legacy endpoints to route_v1_legacy file
