+++ api/pkg/handler/test/hack/hack.go
@@ -53,6 +53,7 @@ func NewTestRouting(
 	mainRouter := mux.NewRouter()
 	v1Router := mainRouter.PathPrefix("/api/v1").Subrouter()
 	r.RegisterV1(v1Router, generateDefaultMetrics())
+	r.RegisterV1Legacy(v1Router)
 	r.RegisterV1Optional(v1Router,
 		true,
 		*generateDefaultOicdCfg(),
